+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+# Changelog\
 
 All notable changes to MarkRead will be documented in this file.
 
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Planned
 - Export to PDF
 - Bookmarks
+
+## [0.6.1] - 2026-01-22
+- improved dark theme
+- enable html markup in the md files (purified)
+- last version changes in the about box
+- improved mermaid  erd diagram coloring
+- markdown footnotes
+- markdown containers (info, warning, error)
 
 ## [0.5.16] - 2026-01-17